Changed param setting method in Const module.
This module now scales and discards any values beyond 2 digits of precision in the float output,
and stores its value as effectively an int to avoid corruption in lower digits, as it's only
possible to manually set the upper two anyway. This means the float output is now only in hundredths
at its smallest range of output.