Reintroduced variable jump length for cursor on PAGE-UP/DOWN. Merged in refactoring from synthgrid-keys branch.
Now checks for cursor being in a zone where PAGE-UP/DOWN respectively will require a jump less than 16
to stay within bounds, then calculates the number of rows needed for jump.

I spent ages trying to figure this out using modulo and so forth, ended up making a while loop in my
previous method to do this, and all the time didn't realise the simplest way was so obvious. Bah.